fixup! [move] Add deprecation attributes to source language
