+++ external-crates/move/crates/move-compiler/src/shared/program_info.rs
@@ -1,7 +1,7 @@
 // Copyright (c) The Move Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{collections::BTreeMap, sync::Arc};
+use std::{collections::BTreeMap, fmt::Display, sync::Arc};
 
 use move_ir_types::location::Loc;
 use move_symbol_pool::Symbol;
@@ -18,6 +18,8 @@ use crate::{
     typing::ast::{self as T},
     FullyCompiledProgram,
 };
+
+use self::known_attributes::AttributePosition;
 
 #[derive(Debug, Clone)]
 pub struct FunctionInfo {
@@ -54,6 +56,14 @@ pub struct ModuleInfo {
 pub enum DatatypeKind {
     Struct,
     Enum,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum NamedMemberKind {
+    Struct,
+    Enum,
+    Function,
+    Constant,
 }
 
 #[derive(Debug, Clone)]
@@ -230,14 +240,26 @@ impl<const AFTER_TYPING: bool> ProgramInfo<AFTER_TYPING> {
         &self.enum_definition(m, n).type_parameters
     }
 
+    pub fn named_member_kind(&self, m: ModuleIdent, n: Name) -> NamedMemberKind {
+        let minfo = self.module(&m);
+        if minfo.structs.contains_key(&DatatypeName(n)) {
+            NamedMemberKind::Struct
+        } else if minfo.enums.contains_key(&DatatypeName(n)) {
+            NamedMemberKind::Enum
+        } else if minfo.functions.contains_key(&FunctionName(n)) {
+            NamedMemberKind::Function
+        } else if minfo.constants.contains_key(&ConstantName(n)) {
+            NamedMemberKind::Constant
+        } else {
+            panic!("ICE should have failed in naming")
+        }
+    }
+
     pub fn datatype_kind(&self, m: &ModuleIdent, n: &DatatypeName) -> DatatypeKind {
-        match (
-            self.module(m).structs.contains_key(n),
-            self.module(m).enums.contains_key(n),
-        ) {
-            (true, false) => DatatypeKind::Struct,
-            (false, true) => DatatypeKind::Enum,
-            (false, false) | (true, true) => panic!("ICE should have failed in naming"),
+        match self.named_member_kind(*m, n.0) {
+            NamedMemberKind::Struct => DatatypeKind::Struct,
+            NamedMemberKind::Enum => DatatypeKind::Enum,
+            _ => panic!("ICE should have failed in naming"),
         }
     }
 
@@ -349,6 +371,37 @@ impl<const AFTER_TYPING: bool> ProgramInfo<AFTER_TYPING> {
         match &vdef.fields {
             N::VariantFields::Empty => false,
             N::VariantFields::Defined(is_positional, _m) => *is_positional,
+        }
+    }
+}
+
+impl Display for NamedMemberKind {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            NamedMemberKind::Struct => write!(f, "struct"),
+            NamedMemberKind::Enum => write!(f, "enum"),
+            NamedMemberKind::Function => write!(f, "function"),
+            NamedMemberKind::Constant => write!(f, "constant"),
+        }
+    }
+}
+
+impl From<NamedMemberKind> for AttributePosition {
+    fn from(nmk: NamedMemberKind) -> Self {
+        match nmk {
+            NamedMemberKind::Struct => AttributePosition::Struct,
+            NamedMemberKind::Enum => AttributePosition::Enum,
+            NamedMemberKind::Function => AttributePosition::Function,
+            NamedMemberKind::Constant => AttributePosition::Constant,
+        }
+    }
+}
+
+impl From<DatatypeKind> for NamedMemberKind {
+    fn from(dt: DatatypeKind) -> Self {
+        match dt {
+            DatatypeKind::Struct => NamedMemberKind::Struct,
+            DatatypeKind::Enum => NamedMemberKind::Enum,
         }
     }
 }
