+++ external-crates/move/crates/move-compiler/tests/move_2024/typing/deprecated_macros.exp
@@ -1,94 +1,58 @@
-warning[W04039]: use of deprecated struct
-   ┌─ tests/move_2024/typing/deprecated_macros.move:8:37
+warning[W04037]: deprecated usage
+   ┌─ tests/move_2024/typing/deprecated_macros.move:10:37
    │
- 8 │     macro fun call_with_x($f: || -> X): X {
-   │                                     ^   - Deprecated struct used here
+10 │     macro fun call_with_x($f: || -> X): X {
+   │                                     ^   - Deprecated struct also used here
    │                                     │    
    │                                     This struct is deprecated
    ·
-25 │         call_with_x!(|| X());
-   │                         - Deprecated struct used here
+27 │         call_with_x!(|| X());
+   │                         - Deprecated struct also used here
 
-warning[W04039]: use of deprecated struct
-  ┌─ tests/move_2024/typing/deprecated_macros.move:8:37
-  │
-8 │     macro fun call_with_x($f: || -> X): X {
-  │                                     ^   - Deprecated struct used here
-  │                                     │    
-  │                                     This struct is deprecated
-
-warning[W04041]: use of deprecated function
-   ┌─ tests/move_2024/typing/deprecated_macros.move:9:9
+warning[W04037]: deprecated usage
+   ┌─ tests/move_2024/typing/deprecated_macros.move:11:9
    │
- 9 │         f();
+11 │         f();
    │         ^ This function is deprecated
    ·
-24 │         f();
-   │         - Deprecated function used here
+26 │         f();
+   │         - Deprecated function also used here
 
-warning[W04041]: use of deprecated function
-   ┌─ tests/move_2024/typing/deprecated_macros.move:14:9
+warning[W04037]: deprecated usage
+   ┌─ tests/move_2024/typing/deprecated_macros.move:16:9
    │
-14 │         f();
+16 │         f();
    │         ^ This function is deprecated
    ·
-29 │         f();
-   │         - Deprecated function used here
+31 │         f();
+   │         - Deprecated function also used here
 
-warning[W04041]: use of deprecated function
-   ┌─ tests/move_2024/typing/deprecated_macros.move:19:9
+warning[W04037]: deprecated usage
+   ┌─ tests/move_2024/typing/deprecated_macros.move:21:9
    │
-19 │         f();
+21 │         f();
    │         ^ This function is deprecated
    ·
-36 │         f();
-   │         - Deprecated function used here
+38 │         f();
+   │         - Deprecated function also used here
 
-warning[W04039]: use of deprecated struct
-   ┌─ tests/move_2024/typing/deprecated_macros.move:30:18
-   │
-30 │         call!(|| X());
-   │                  ^^^ This struct is deprecated
-
-error[E03005]: unbound unscoped name
-   ┌─ tests/move_2024/typing/deprecated_macros.move:31:18
-   │
-31 │         call!(|| Y<X>());
-   │                  ^ Unbound function 'Y' in current scope
-
-error[E03005]: unbound unscoped name
+warning[W04037]: deprecated usage
    ┌─ tests/move_2024/typing/deprecated_macros.move:32:18
    │
-32 │         call!(|| Y<Y<X>>());
-   │                  ^ Unbound function 'Y' in current scope
+32 │         call!(|| X());
+   │                  ^ This struct is deprecated
+33 │         call!(|| Y<X>());
+   │                    - Deprecated struct also used here
+34 │         call!(|| Y<Y<X>>());
+   │                      - Deprecated struct also used here
 
-error[E03004]: unbound type
-   ┌─ tests/move_2024/typing/deprecated_macros.move:32:20
-   │
-32 │         call!(|| Y<Y<X>>());
-   │                    ^ Unbound type 'Y' in current scope
-
-warning[W04039]: use of deprecated struct
-   ┌─ tests/move_2024/typing/deprecated_macros.move:37:25
-   │
-37 │         call_unders!(|| X());
-   │                         ^^^ This struct is deprecated
-
-error[E03005]: unbound unscoped name
-   ┌─ tests/move_2024/typing/deprecated_macros.move:38:25
-   │
-38 │         call_unders!(|| Y<X>());
-   │                         ^ Unbound function 'Y' in current scope
-
-error[E03005]: unbound unscoped name
+warning[W04037]: deprecated usage
    ┌─ tests/move_2024/typing/deprecated_macros.move:39:25
    │
-39 │         call_unders!(|| Y<Y<X>>());
-   │                         ^ Unbound function 'Y' in current scope
-
-error[E03004]: unbound type
-   ┌─ tests/move_2024/typing/deprecated_macros.move:39:27
-   │
-39 │         call_unders!(|| Y<Y<X>>());
-   │                           ^ Unbound type 'Y' in current scope
+39 │         call_unders!(|| X());
+   │                         ^ This struct is deprecated
+40 │         call_unders!(|| Y<X>());
+   │                           - Deprecated struct also used here
+41 │         call_unders!(|| Y<Y<X>>());
+   │                             - Deprecated struct also used here
