+++ external-crates/move/crates/move-compiler/tests/move_2024/typing/deprecated_macros.move
@@ -2,6 +2,8 @@ module 0x42::m {
     #[deprecated]
     public struct X() has drop;
 
+    public struct Y<phantom T>() has drop;
+
     #[deprecated]
     fun f() { }
