+++ external-crates/move/crates/move-compiler/tests/move_2024/typing/deprecation_inter_module_calls.exp
@@ -1,61 +1,61 @@
-warning[W04038]: use of deprecated module
+warning[W04037]: deprecated usage
    ┌─ tests/move_2024/typing/deprecation_inter_module_calls.move:23:12
    │
 23 │         f: X,
-   │            ^ This struct is deprecated since the whole module that it is declared in is marked deprecated: Use the k module instead.
+   │            ^ This struct is deprecated. It is deprecated since its whole module is marked deprecated: Use the k module instead.
 
-warning[W04038]: use of deprecated module
+warning[W04037]: deprecated usage
    ┌─ tests/move_2024/typing/deprecation_inter_module_calls.move:27:14
    │
 27 │         Some(X),
-   │              ^ This struct is deprecated since the whole module that it is declared in is marked deprecated: Use the k module instead.
+   │              ^ This struct is deprecated. It is deprecated since its whole module is marked deprecated: Use the k module instead.
 28 │         Other { y: X },
-   │                    - Deprecated struct used here
+   │                    - Deprecated struct also used here
 
-warning[W04038]: use of deprecated module
+warning[W04037]: deprecated usage
    ┌─ tests/move_2024/typing/deprecation_inter_module_calls.move:38:24
    │
 38 │     public fun lol(_: &X) { }
-   │                        ^ This struct is deprecated since the whole module that it is declared in is marked deprecated: Use the k module instead.
+   │                        ^ This struct is deprecated. It is deprecated since its whole module is marked deprecated: Use the k module instead.
 
-warning[W04038]: use of deprecated module
-   ┌─ tests/move_2024/typing/deprecation_inter_module_calls.move:40:24
+warning[W04037]: deprecated usage
+   ┌─ tests/move_2024/typing/deprecation_inter_module_calls.move:40:32
    │
 40 │     public fun quux(x: 0x7::l::X) {
-   │                        ^^^^^^^^^ This struct is deprecated since the whole module that it is declared in is marked deprecated: Use the k module instead.
+   │                                ^ This struct is deprecated. It is deprecated since its whole module is marked deprecated: Use the k module instead.
 41 │         0x7::l::foo();
-   │                 --- Deprecated function used here
+   │                 --- Deprecated function also used here
 42 │         x.bar();
-   │           --- Deprecated function used here
+   │           --- Deprecated function also used here
 
-warning[W04038]: use of deprecated module
+warning[W04037]: deprecated usage
    ┌─ tests/move_2024/typing/deprecation_inter_module_calls.move:47:12
    │
 47 │         l::foo(); 
-   │            ^^^ This function is deprecated since the whole module that it is declared in is marked deprecated: Use the k module instead.
+   │            ^^^ This function is deprecated. It is deprecated since its whole module is marked deprecated: Use the k module instead.
 
-warning[W04038]: use of deprecated module
+warning[W04037]: deprecated usage
    ┌─ tests/move_2024/typing/deprecation_inter_module_calls.move:52:34
    │
 52 │     public fun dep_meth_call(x: &X) {
-   │                                  ^ This struct is deprecated since the whole module that it is declared in is marked deprecated: Use the k module instead.
+   │                                  ^ This struct is deprecated. It is deprecated since its whole module is marked deprecated: Use the k module instead.
 53 │         x.bar();
-   │           --- Deprecated function used here
+   │           --- Deprecated function also used here
    ·
 59 │         x.rofl();
-   │           ---- Deprecated function used here
+   │           ---- Deprecated function also used here
 
-warning[W04041]: use of deprecated function
+warning[W04037]: deprecated usage
    ┌─ tests/move_2024/typing/deprecation_inter_module_calls.move:58:11
    │
 58 │         x.lol();
    │           ^^^ This function is deprecated: More specific deprecations override less specific ones.
 
-warning[W04038]: use of deprecated module
+warning[W04037]: deprecated usage
    ┌─ tests/move_2024/typing/deprecation_inter_module_calls.move:66:33
    │
 66 │         let _ = T::Some(0x7::l::make_x());
-   │                                 ^^^^^^ This function is deprecated since the whole module that it is declared in is marked deprecated: Use the k module instead.
+   │                                 ^^^^^^ This function is deprecated. It is deprecated since its whole module is marked deprecated: Use the k module instead.
 67 │         let _ = T::Other { y: 0x7::l::make_x()};
-   │                                       ------ Deprecated function used here
+   │                                       ------ Deprecated function also used here
