+++ external-crates/move/crates/move-compiler/tests/move_2024/typing/internal_module_deprecations.exp
@@ -1,70 +1,72 @@
-warning[W04039]: use of deprecated struct
+warning[W04037]: deprecated usage
   ┌─ tests/move_2024/typing/internal_module_deprecations.move:3:13
   │
 3 │         Bad(A),
   │             ^ This struct is deprecated: Use the B struct instead.
 
-warning[W04041]: use of deprecated function
+warning[W04037]: deprecated usage
    ┌─ tests/move_2024/typing/internal_module_deprecations.move:26:9
    │
 26 │         foo();
    │         ^^^ This function is deprecated
+27 │         foo();
+   │         --- Deprecated function also used here
 
-warning[W04041]: use of deprecated function
-   ┌─ tests/move_2024/typing/internal_module_deprecations.move:27:9
+warning[W04037]: deprecated usage
+   ┌─ tests/move_2024/typing/internal_module_deprecations.move:28:9
    │
-27 │         bar_dep();
+28 │         bar_dep();
    │         ^^^^^^^ This function is deprecated: Use the baz function instead.
 
-warning[W04039]: use of deprecated struct
-   ┌─ tests/move_2024/typing/internal_module_deprecations.move:30:23
+warning[W04037]: deprecated usage
+   ┌─ tests/move_2024/typing/internal_module_deprecations.move:31:23
    │
-30 │     public fun qux(_: A) { }
+31 │     public fun qux(_: A) { }
    │                       ^ This struct is deprecated: Use the B struct instead.
 
-warning[W04039]: use of deprecated struct
-   ┌─ tests/move_2024/typing/internal_module_deprecations.move:32:30
+warning[W04037]: deprecated usage
+   ┌─ tests/move_2024/typing/internal_module_deprecations.move:33:30
    │
-32 │     public fun return_dep(): A {
+33 │     public fun return_dep(): A {
    │                              ^ This struct is deprecated: Use the B struct instead.
-33 │         A()
-   │         - Deprecated struct used here
+34 │         A()
+   │         - Deprecated struct also used here
 
-warning[W04037]: use of deprecated constant
-   ┌─ tests/move_2024/typing/internal_module_deprecations.move:37:9
+warning[W04037]: deprecated usage
+   ┌─ tests/move_2024/typing/internal_module_deprecations.move:38:9
    │
-37 │         H
+38 │         H
    │         ^ This constant is deprecated: Use the L constant instead.
 
-warning[W04037]: use of deprecated constant
-   ┌─ tests/move_2024/typing/internal_module_deprecations.move:41:15
+warning[W04037]: deprecated usage
+   ┌─ tests/move_2024/typing/internal_module_deprecations.move:42:15
    │
-41 │         abort OldError
+42 │         abort OldError
    │               ^^^^^^^^ This constant is deprecated: Use `NewError` instead.
 
-warning[W04039]: use of deprecated struct
-   ┌─ tests/move_2024/typing/internal_module_deprecations.move:45:18
+warning[W04037]: deprecated usage
+   ┌─ tests/move_2024/typing/internal_module_deprecations.move:46:18
    │
-45 │         let y = (A(): A);
-   │                  ^    - Deprecated struct used here
+46 │         let y = (A(): A);
+   │                  ^    - Deprecated struct also used here
    │                  │     
    │                  This struct is deprecated: Use the B struct instead.
-46 │         match (y) {
-47 │             A() => (),
-   │             - Deprecated struct used here
-48 │         };
-49 │         A() = A();
-   │         ---   - Deprecated struct used here
+47 │         match (y) {
+48 │             A() => (),
+   │             - Deprecated struct also used here
+49 │         };
+50 │         A() = A();
+   │         -     - Deprecated struct also used here
    │         │      
-   │         Deprecated struct used here
-50 │         let A() = A();
-   │             ---   - Deprecated struct used here
+   │         Deprecated struct also used here
+51 │         let A() = A();
+   │             -     - Deprecated struct also used here
    │             │      
-   │             Deprecated struct used here
+   │             Deprecated struct also used here
 
-warning[W04039]: use of deprecated struct
-   ┌─ tests/move_2024/typing/internal_module_deprecations.move:54:24
+warning[W04037]: deprecated usage
+   ┌─ tests/move_2024/typing/internal_module_deprecations.move:55:24
    │
-54 │         let _ = R::Bad(A());
+55 │         let _ = R::Bad(A());
    │                        ^ This struct is deprecated: Use the B struct instead.
