+++ external-crates/move/crates/move-compiler/tests/move_2024/typing/internal_module_deprecations.move
@@ -24,6 +24,7 @@ module 0x42::m {
 
     public fun baz() { 
         foo();
+        foo();
         bar_dep();
     }
